fix dashboard 404 and nodes uptime parse error
Dashboard was hitting /api/system/stats/system-stats which doesn't exist;
correct path is /api/system/stats. Node and dashboard uptime fields changed
from Long to Double in DTOs — the backend returns fractional seconds (e.g.
2338592.49); domain models stay Long via .toLong().

+++ composeApp/src/commonMain/kotlin/art/yniyniyni/freedomwave/data/api/dto/NodeDto.kt
@@ -57,6 +57,6 @@ data class NodeSystemInfoDto(
 data class NodeSystemStatsDto(
     @SerialName("memoryFree") val memoryFree: Long,
     @SerialName("memoryUsed") val memoryUsed: Long,
-    @SerialName("uptime")     val uptime: Long,
+    @SerialName("uptime")     val uptime: Double,
     @SerialName("loadAvg")    val loadAvg: List<Double> = emptyList()
 )
